nova-debate-arena/requirements.txt
# Core LLM + agent orchestration
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0

# API server + realtime UI
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Data + env
pandas>=2.0.0
python-dotenv>=1.0.0

# Structured outputs / typing
pydantic>=2.0.0
